+++ packages/browser-wallet/src/background/web3Id.ts
@@ -14,6 +14,7 @@ import {
     AttributeType,
     isTimestampAttribute,
     TimestampAttribute,
+    AttributeKeyString,
 } from '@concordium/web-sdk';
 import {
     sessionVerifiableCredentials,
@@ -42,6 +43,8 @@ import {
 import { getNet } from '@shared/utils/network-helpers';
 import { WAIT_FOR_CLOSED_POPUP_ITERATIONS, WAIT_FOR_CLOSED_POPUP_TIMEOUT_MS } from '@shared/constants/web3id';
 import { Buffer } from 'buffer/';
+import { isAgeStatement, SecretStatement } from '@popup/pages/IdProofRequest/DisplayStatement/utils';
+import { logError } from '@shared/utils/log-helpers';
 import { openWindow, RunCondition, testPopupOpen } from './window-management';
 import bgMessageHandler from './message-handler';
 
@@ -246,8 +249,8 @@ export const createWeb3IdProofHandler: ExtensionMessageHandler = (msg, _sender, 
 };
 
 export const runIfValidWeb3IdProof: RunCondition<MessageStatusWrapper<undefined>> = async (msg) => {
-    if (!isHex(msg.payload.challenge)) {
-        return rejectRequest(`Challenge is invalid, it should be a HEX encoded string`);
+    if (!isHex(msg.payload.challenge) || msg.payload.challenge.length !== 64) {
+        return rejectRequest(`Challenge is invalid, it should be 32 bytes as a HEX encoded string`);
     }
     try {
         const statements: CredentialStatements = parse(msg.payload.statements);
@@ -343,3 +346,31 @@ export const loadWeb3IdBackupHandler: ExtensionMessageHandler = (_msg, _sender, 
     loadWeb3IdBackup();
     respond(undefined);
 };
+
+/**
+ * Check if the web3IdProof payload statements are an "Age proof"
+ * i.e. it is a single account credential statement with a single atomic statement, which is an age statement.
+ * n.b. We have this to filter some request to a special page for age statements.
+ */
+export function isAgeProof(payload: { statements: string }): boolean {
+    try {
+        const statements: CredentialStatements = parse(payload.statements);
+        const credentialStatement = statements[0];
+
+        if (!(statements.length === 1 && isAccountCredentialStatement(credentialStatement))) {
+            return false;
+        }
+
+        const atomicStatement = credentialStatement.statement[0];
+
+        return (
+            credentialStatement.statement.length === 1 &&
+            atomicStatement.type === StatementTypes.AttributeInRange &&
+            atomicStatement.attributeTag === AttributeKeyString.dob &&
+            isAgeStatement(atomicStatement as SecretStatement)
+        );
+    } catch (e) {
+        logError(e);
+        return false;
+    }
+}