+++ packages/browser-wallet/src/popup/constants/routes.ts
@@ -75,6 +75,9 @@ export const relativeRoutes = {
         web3IdProof: {
             path: 'web3Id-proof',
         },
+        ageProof: {
+            path: 'age-proof',
+        },
     },
     setup: {
         path: '/setup',